feat: search state management and cleanup
