--- packages/app/src/app/components/CreateNewSandbox/CreateSandbox/Explore/Search/index.ts
@@ -1,1 +1,0 @@
-export { ExploreSearch } from './ExploreSearch';
