--- packages/app/src/app/components/CreateNewSandbox/CreateSandbox/Explore/SearchResults/elements.ts
@@ -1,9 +1,0 @@
-import { createGlobalStyle } from 'styled-components';
-
-export const GlobalSearchStyles = createGlobalStyle`
-.ais-InstantSearch__root {
-  display: flex;
-  flex-direction: column;
-  height: 496px;
-}
-`;
