--- packages/app/src/app/components/CreateNewSandbox/CreateSandbox/Explore/index.ts
@@ -1,1 +1,0 @@
-export { Explore } from './Explore';
