--- packages/app/src/app/components/CreateNewSandbox/CreateSandbox/Explore/utils.ts
@@ -1,16 +1,0 @@
-import { useState, useEffect } from 'react';
-
-export function useDebounce(value: string | number, delay: number) {
-  const [debouncedValue, setDebouncedValue] = useState(value);
-
-  useEffect(() => {
-    const handler = setTimeout(() => {
-      setDebouncedValue(value);
-    }, delay);
-    return () => {
-      clearTimeout(handler);
-    };
-  }, [delay, value]);
-
-  return debouncedValue;
-}
