--- packages/app/src/app/components/CreateNewSandbox/CreateSandbox/QuickStart/PersonalTemplates/PersonalTemplates.tsx
@@ -1,22 +1,0 @@
-import React from 'react';
-import { ITemplateInfo } from '../../TemplateList';
-import { DynamicWidthTemplateList } from '../../TemplateList/DynamicWidthTemplateList';
-
-interface IPersonalTemplatesProps {
-  officialTemplates: ITemplateInfo[];
-  collectionId?: string;
-}
-
-export const PersonalTemplates = ({
-  officialTemplates,
-  collectionId,
-}: IPersonalTemplatesProps) => {
-  return (
-    <DynamicWidthTemplateList
-      showSecondaryShortcuts
-      forkOnOpen
-      templateInfos={officialTemplates}
-      collectionId={collectionId}
-    />
-  );
-};
