--- packages/app/src/app/components/CreateNewSandbox/CreateSandbox/QuickStart/PersonalTemplates/index.ts
@@ -1,1 +1,0 @@
-export { PersonalTemplates } from './PersonalTemplates';
