--- packages/app/src/app/components/CreateNewSandbox/CreateSandbox/QuickStart/index.ts
@@ -1,1 +1,0 @@
-export { QuickStart } from './QuickStart';
