--- packages/app/src/app/components/CreateNewSandbox/CreateSandbox/TemplateList/index.ts
@@ -1,1 +1,0 @@
-export { TemplateList, ITemplateInfo } from './TemplateList';
